display timecode from BWF etc in SF browser; accels for windows now work (misspelling in ardour.bindings); use hide icon in automation tracks
git-svn-id: svn://localhost/ardour2/trunk@1125 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ libs/pbd/pbd/convert.h
@@ -31,7 +31,8 @@ int    atoi (const std::string&);
 double atof (const std::string&);
 void   url_decode (std::string&);
 
-std::string length2string (const int32_t frames, const float sample_rate);
+// std::string length2string (const int32_t frames, const float sample_rate);
+std::string length2string (const int64_t frames, const double sample_rate);
 
 std::vector<std::string> internationalize (const char *, const char **);
